// packages/core/src/types/bot.d.ts
declare namespace BotModel {
  type BotPair = {
    id: number;
    pair_id: string;
    name: string;
    base_token: Token;
    quote_token: Token;
    symbol: string;
    ticker: PairTicker;
    is_main: boolean;
    is_meme: boolean;
    chain: Chain;
    support_grid?: boolean;
    support_dca?: boolean;
  };

  type PairTicker = {
    pair_id: string;
    last_price: string;
    high: string;
    low: string;
    open: string;
    quote_volume: string;
    quote_usd: string;
    price_change_percent: string;
    all_quote_volume: string;
    all_quote_usd: string;
  };

  type BotType = 'grid' | 'swing' | 'dca';
  type GridBotType = 'grid' | 'swing';
  type BotStatus = 'active' | 'expired' | 'closed';

  interface Token {
    id: number;
    code: string;
    symbol: string;
    name: string;
    icon: string;
    decimals: number;
    oracle_id: string;
    volume: string;
    amount?: string;
  }
// ...
}